Migrate to sprotty ToolManager
