+++ client/packages/glsp-sprotty/package.json
@@ -8,7 +8,7 @@
   ],
   "dependencies": {
     "autocompleter": "^4.0.2",
-    "sprotty": "next"
+    "sprotty": "0.5.0-next.d035414"
   },
   "devDependencies": {
     "rimraf": "latest",
