+++ client/packages/glsp-sprotty/src/features/tools/default-tools.ts
@@ -14,13 +14,12 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  ********************************************************************************/
 import { interfaces } from "inversify";
-import { ToolManager } from "../../base/tool-manager/tool-manager";
-import { GLSP_TYPES } from "../../types";
+import { ToolManager, TYPES } from "sprotty/lib";
 import { ChangeBoundsTool } from "./change-bounds-tool";
 import { DelKeyDeleteTool, MouseDeleteTool } from "./delete-tool";
 
 export function registerDefaultTools(container: interfaces.Container) {
-    const toolManager: ToolManager = container.get(GLSP_TYPES.IToolManager);
+    const toolManager: ToolManager = container.get(TYPES.IToolManager);
     toolManager.registerDefaultTools(
         container.resolve(ChangeBoundsTool),
         container.resolve(DelKeyDeleteTool));
