+++ client/packages/glsp-sprotty/src/index.ts
@@ -18,8 +18,7 @@ export * from './base/command-stack';
 export * from './base/diagram-ui-extension/diagram-ui-extension';
 export * from './base/edit-config/edit-config';
 export * from './base/model/model-update-observer-registry';
-export * from './base/tool-manager/tool';
-export * from './base/tool-manager/tool-manager';
+export * from './base/tool-manager/tool-manager-action-handler';
 export * from './features/change-bounds/model';
 export * from './features/change-bounds/resize';
 export * from './features/command-palette/action-provider';
