+++ client/packages/glsp-sprotty/src/types.ts
@@ -14,7 +14,6 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  ********************************************************************************/
 export const GLSP_TYPES = {
-    IToolManager: Symbol.for("ToolManager"),
     ICommandPaletteActionProvider: Symbol.for("ICommandPaletteActionProvider"),
     ICommandPaletteActionProviderRegistry: Symbol.for("ICommandPaletteActionProviderRegistry"),
     IFeedbackActionDispatcher: Symbol.for("IFeedbackActionDispatcher"),
